+++ src/admin/components/elements/Button/index.tsx
@@ -1,4 +1,4 @@
-import React, { Fragment, isValidElement } from 'react';
+import React, { forwardRef, Fragment, isValidElement } from 'react';
 import { Link } from 'react-router-dom';
 import { Props } from './types';
 
@@ -53,7 +53,7 @@ const ButtonContents = ({ children, icon, tooltip, showTooltip }) => {
   );
 };
 
-const Button: React.FC<Props> = (props) => {
+const Button = forwardRef<HTMLButtonElement | HTMLAnchorElement, Props>((props, ref) => {
   const {
     className,
     id,
@@ -129,6 +129,7 @@ const Button: React.FC<Props> = (props) => {
       return (
         <a
           {...buttonProps}
+          ref={ref as React.LegacyRef<HTMLAnchorElement>}
           href={url}
         >
           <ButtonContents
@@ -147,6 +148,7 @@ const Button: React.FC<Props> = (props) => {
       return (
         <Tag
           type="submit"
+          ref={ref}
           {...buttonProps}
         >
           <ButtonContents
@@ -159,6 +161,6 @@ const Button: React.FC<Props> = (props) => {
         </Tag>
       );
   }
-};
+});
 
 export default Button;