+++ src/admin/components/forms/Form/index.tsx
@@ -28,7 +28,7 @@ import { useOperation } from '../../utilities/OperationProvider';
 import { WatchFormErrors } from './WatchFormErrors';
 import { splitPathByArrayFields } from '../../../../utilities/splitPathByArrayFields';
 import { setsAreEqual } from '../../../../utilities/setsAreEqual';
-import { buildFieldSchemaMap } from '../../../../utilities/buildFieldSchemaMap';
+import { buildFieldSchemaMap } from './buildFieldSchemaMap';
 import { isNumber } from '../../../../utilities/isNumber';
 
 const baseClass = 'form';
@@ -426,7 +426,7 @@ const Form: React.FC<Props> = (props) => {
 
     if (fieldConfig) {
       const subFieldState = await buildStateFromSchema({ fieldSchema: fieldConfig, data, preferences, operation, id, user, locale, t });
-      dispatchFields({ type: 'ADD_ROW', rowIndex, path, blockType: data?.blockType, subFieldState });
+      dispatchFields({ type: 'ADD_ROW', rowIndex: rowIndex - 1, path, blockType: data?.blockType, subFieldState });
     }
   }, [dispatchFields, getDocPreferences, id, user, operation, locale, t, getRowConfigByPath]);
 
@@ -443,7 +443,7 @@ const Form: React.FC<Props> = (props) => {
 
     if (fieldConfig) {
       const subFieldState = await buildStateFromSchema({ fieldSchema: fieldConfig, data, preferences, operation, id, user, locale, t });
-      dispatchFields({ type: 'REPLACE_ROW', rowIndex, path, blockType: data?.blockType, subFieldState });
+      dispatchFields({ type: 'REPLACE_ROW', rowIndex: rowIndex - 1, path, blockType: data?.blockType, subFieldState });
     }
   }, [dispatchFields, getDocPreferences, id, user, operation, locale, t, getRowConfigByPath]);