+++ src/admin/components/forms/field-types/RichText/elements/li/index.tsx
@@ -7,7 +7,7 @@ const LI = (props) => {
 
   return (
     <li
-      style={{ listStyle: disableListStyle ? 'none' : undefined }}
+      style={{ listStyle: disableListStyle ? 'none' : undefined, listStylePosition: disableListStyle ? 'outside' : undefined }}
       {...attributes}
     >
       {children}